Remove self reactions
