+++ src/index.js
@@ -9,6 +9,7 @@ const {
 const { environment } = require('./environment')
 const { error, log } = require('./utils')
 const parseWhitelistedChannels = require('./parser/whitelistedChannels')
+const roleMessage = require('./roleMessage')
 
 // Load this as early as possible, to init all the environment variables that may be needed
 dotenv.config()
@@ -18,32 +19,16 @@ const client = new Discord.Client()
 
 client.on('ready', async () => {
   log(`Bot successfully started as ${client.user.tag}`)
-  const filter = (reaction) => {
-    return (
-      reaction.emoji.name === '💻' ||
-      reaction.emoji.name === '🎨' ||
-      reaction.emoji.name === '🛒'
-    )
+  roleMessage(client)
+})
+
+client.on('messageReactionAdd', (reaction, user) => {
+  let message = reaction.message
+  if (message.author.id === user.id) {
+    // Remove the user's reaction
+    reaction.remove(user)
+    log(`Removed self reaction for ${user.tag}`)
   }
-  const message = await client.channels.cache
-    .get('765653777691836428')
-    .messages.fetch('765662524900769872')
-  const collector = message.createReactionCollector(filter)
-  collector.on('collect', (reaction, user) => {
-    const member = message.guild.members.cache.get(user.id)
-    if (reaction.emoji.name === '💻') {
-      const role = message.guild.roles.cache.find(r => r.name === 'Developer')
-      member.roles.add(role)
-    } else if (reaction.emoji.name === '🎨') {
-      const role = message.guild.roles.cache.find(r => r.name === 'Designer')
-      member.roles.add(role)
-    } else if (reaction.emoji.name === '🛒') {
-      const role = message.guild.roles.cache.find(
-        r => r.name === 'Social Media Unicorn',
-      )
-      member.roles.add(role)
-    } else return null
-  })
 })
 
 client.on('message', message => {
